SABADBAN-585 8 fix some bugs. change symbol code to Int64 everywhere except database.

+++ trunk/SabadBan/SabadBan/AddAlarmResponse.swift
@@ -22,7 +22,7 @@ public struct AddAlarmResponse: Glossy {
     public let id: Int!
     public let orderType: String!
     public let priceDirection: String!
-    public let symbolCode: Int!
+    public let symbolCode: Int64!
 
 
     //MARK: Decodable
